Moved PHMessageHandler to wsconnector.py
